requirements.txt
gTTS
pydub
openai==0.27.8
speechrecognition
pyaudio
datetime
python-dotenv
torch
whisper
soundfile
eel
screeninfo
python-dotenv
pyaudio
numpy